it's annoying to have to put stupid casts when passing in constant numbers (which are ints or longs or something in Java ...)

+++ src/com/blarg/gdx/tilemap3d/Tile.java
@@ -42,24 +42,24 @@ public final class Tile {
 		tile = NO_TILE;
 	}
 
-	public Tile set(short tileIndex) {
-		this.tile = tileIndex;
+	public Tile set(int tileIndex) {
+		this.tile = (short)tileIndex;
 		return this;
 	}
 
-	public Tile set(short tileIndex, short flags) {
-		this.tile = tileIndex;
+	public Tile set(int tileIndex, short flags) {
+		this.tile = (short)tileIndex;
 		this.flags = flags;
 		return this;
 	}
 
-	public Tile set(short tileIndex, short flags, final Color color) {
+	public Tile set(int tileIndex, short flags, final Color color) {
 		return set(tileIndex, flags, color.toIntBits());
 	}
 
-	public Tile set(short tileIndex, short flags, int color) {
+	public Tile set(int tileIndex, short flags, int color) {
 		flags = Bitfield.set(FLAG_CUSTOM_COLOR, flags);
-		this.tile = tileIndex;
+		this.tile = (short)tileIndex;
 		this.flags = flags;
 		this.color = color;
 		return this;
